+++ ggshow/ggshow.py
@@ -38,7 +38,7 @@ if not _find_rscript():
 def ggwrite(plotcode: str, outfile: str, libs: tuple=(),
             savesize: tuple=None, width: float=None, height: float=None,
             scale: float=1, units: str="in", dpi: int=300,
-            rscriptpath: str=None, message_encoding: str="utf-8", **data):
+            rscriptcommand: str=None, message_encoding: str="utf-8", **data):
     """
     Write a ggplot2 graph to a file
 
@@ -52,7 +52,7 @@ def ggwrite(plotcode: str, outfile: str, libs: tuple=(),
         scale            :   ggsave option scale
         units            :   ggsave option units
         dpi              :   ggsave option gpi
-        rscriptpath      :   If given, used as the path to the Rscript command
+        rscriptcommand   :   If given, used as the Rscript command name
         message_encoding :   Character encoding of the subprocess outputs
         **data           :   pandas data frames with names used inside the R script
 
@@ -77,8 +77,8 @@ def ggwrite(plotcode: str, outfile: str, libs: tuple=(),
         if width is None: width = "NA"
         if height is None: height = "NA"
         outfile = Path(outfile).as_posix()  # normalize the path string on windowns
-        if rscriptpath is None:
-            rscriptpath = config.rscript
+        if rscriptcommand is None:
+            rscriptcommand = config.rscript
 
         code = """
         {importcode}
@@ -95,7 +95,7 @@ def ggwrite(plotcode: str, outfile: str, libs: tuple=(),
             codefile = os.path.join(tmpdir, "__ggcode.R")
             with open(codefile, "w", encoding="utf-8") as f:
                 f.write(code)
-            p = subprocess.run([rscriptpath, codefile], stdout=subprocess.PIPE, stderr=subprocess.PIPE)        
+            p = subprocess.run([rscriptcommand, codefile], stdout=subprocess.PIPE, stderr=subprocess.PIPE)        
         if p.returncode != 0:
             warnmessage = ("Some error occurred while running the R code. The graph may not have been created."
                            "\nStdout:\n\n{}\nStderr:\n\n{}\nR code (auto-generated):\n{}").format(
@@ -108,7 +108,7 @@ def ggwrite(plotcode: str, outfile: str, libs: tuple=(),
 
 def ggshow(plotcode: str, dispwidth: float=300, dispheight: float=None, libs: tuple=(), imageformat: str="png", display: bool=True,
            savesize: tuple=None, width: float=None, height: float=None, scale: float=1, units: str="in", dpi: int=300,
-           rscriptpath: str=None, message_encoding: str="utf-8", **data)-> Image:
+           rscriptcommand: str=None, message_encoding: str="utf-8", **data)-> Image:
     """
     Draw a ggplot2 graph
 
@@ -123,7 +123,7 @@ def ggshow(plotcode: str, dispwidth: float=300, dispheight: float=None, libs: tu
         scale            :   ggsave option scale
         units            :   ggsave option units
         dpi              :   ggsave option gpi
-        rscriptpath      :   If given, used as the path to the Rscript executable file
+        rscriptcommand   :   If given, used as the the Rscript command name
         message_encoding :   Character encoding of the subprocess outputs
         **data      :   pandas data frames with names used inside the R script
 
@@ -135,7 +135,7 @@ def ggshow(plotcode: str, dispwidth: float=300, dispheight: float=None, libs: tu
         outfile = os.path.join(tmpdir, "__ggout." + imageformat)
         ggwrite(plotcode, outfile, libs=libs,
                 savesize=savesize, width=width, height=height, scale=scale, units=units, dpi=dpi,
-                rscriptpath=rscriptpath, message_encoding=message_encoding, **data)
+                rscriptcommand=rscriptcommand, message_encoding=message_encoding, **data)
         if not os.path.isfile(outfile):
             raise RuntimeError("Graph file not found. Perhaps Rscript failed to produce the graph")
         
@@ -168,6 +168,8 @@ try:
         @argument("--scale", type=float, default=1, help="ggsave option scale")
         @argument("--units", type=str, default="in", help="ggsave option units")
         @argument("--dpi", type=int, default=300, help="ggsave option dpi")
+        @argument("--message_encoding", type=str, default="utf-8", help="Encoding of the subprocess outputs")
+        @argument("--rscriptcommand", type=str, default=None, help="Rscript command name")
         @argument("-w", "--dispwidth", type=float, default=None, help="display width")
         @argument("-h", "--dispheight", type=float, default=None, help="display width")
         @argument("--imageformat", type=str, default="png", choices=("png", "jpeg", "svg"), help="imagefile format")